Refactor `Network` dispatches
